backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ricardo_Li <46476463+RicardoYam@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
sqlalchemy = "^2.0.31"
flask-cors = "^4.0.1"
flask-sqlalchemy = "^3.1.1"
bcrypt = "^4.1.3"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
